Add 'elapse_minutes' property to message model
'elapse_minutes' is needed to track how long a user wants to delay a 
specific message before sending it. For example, if a message should be 
sent 24 hours after subscribing a phone number to a campaign, the 
elapse_minutes on that specific message would be set to 1440. The UI 
will most likely only present the ability to select how many hours to 
wait, but minutes may be useful in the future