Update sms raketask flow to pass around id
The rake task for sending text messages had previously been passing 
around an ActiveRecord contact_message object. When it was added as an 
async task it was dumped into Redis. When the task was retrieved the 
contact_message object came out as a string. I forgor Redis only deals 
with strings. At first I thought about just passing the reciever phone 
number and the message, but then realized we could just pass the 
contact_message id